requirements.txt
django==4.2.9
djangorestframework==3.14.0
django_cors_headers==4.3.1
psycopg2-binary==2.9.9
django-filter==23.5
coreapi==2.3.3
markdown==3.5.2
